+++ src/modules/project/components/ProjectWorkspace.vue
@@ -181,7 +181,7 @@ onMounted(() => {
 .project-workspace {
     display: flex;
     flex-direction: column;
-    height: calc(100vh - 88px);
+    height: calc(100vh - 88px - var(--window-titlebar-height, 0px));
     padding: 1.5rem;
     overflow: hidden;
 }